Extract controller into module
